Change classes to files in the lib folder and add game flow in the main.rb
