+++ bin/main.rb
@@ -1,114 +1,60 @@
 #!/usr/bin/env ruby
-# rubocop:disable Metrics/PerceivedComplexity,Metrics/CyclomaticComplexity
+require_relative '../lib/game'
+require_relative '../lib/board'
+require_relative '../lib/player.rb'
 
-class Game
-  attr_accessor :turn, :board
-  attr_reader :player1, :player2
+while true
+  new_board = Board.new
 
-  def initialize(player1 = Player, player2 = Player, board = Board)
-    @player1 = player1
-    @player2 = player2
-    @board = board
-    @turn = 0
-  end
+  puts 'First player? '
+  first_player = Player.new(gets.chomp.capitalize, 'X')
+  puts 'Second player? '
+  second_player = Player.new(gets.chomp.capitalize, 'O')
 
-  def round
-    @board.display_board
-    @turn += 1
-    win_pos = check_win_move(@player1.token) if @turn.odd?
-    win_pos = check_win_move(@player2.token) if @turn.even?
+  new_game = Game.new(first_player, second_player, new_board)
+
+  # Starts the game
+  while new_game.turn < 9
+    new_board.display_board
+    new_game.turn += 1
+    # Checks for the winning position
+    win_pos = new_game.check_win_move(first_player.token, new_board.board) if new_game.turn.odd?
+    win_pos = new_game.check_win_move(second_player.token, new_board.board) if new_game.turn.even?
     puts 'YOU CAN WIN THIS!!!!!' if win_pos
-    print "#{@player1.name} (#{@player1.token}): " if @turn.odd?
-    print "#{@player2.name} (#{@player2.token}): " if @turn.even?
+
+    # Asks the player for the position
+    print "#{first_player.name} (#{first_player.token}): " if new_game.turn.odd?
+    print "#{second_player.name} (#{second_player.token}): " if new_game.turn.even?
     puts 'Which position to take?'
     input = gets.chomp.to_i
-    if @board.include?(input)
-      @board[input - 1] = @player1.token if @turn.odd?
-      @board[input - 1] = @player2.token if @turn.even?
+
+    # Checks if the position is valid
+    pos_valid=new_game.check_position(input, new_board.board)
+    if pos_valid
+      new_board.take_position(input, first_player.token) if new_game.turn.odd?
+      new_board.take_position(input, second_player.token) if new_game.turn.even?
     else
-      @turn -= 1
+      new_game.turn -= 1
       puts 'Invalid play! Try again!'
     end
-  end
 
-  def start_game
-    while @turn < 9
-      round
-      win = check_victory(@player1.token) if @turn.odd?
-      win = check_victory(@player2.token) if @turn.even?
-      finish_game if win
+    #Victory check
+    win = new_game.check_victory(first_player.token, new_board.board) if new_game.turn.odd?
+    win = new_game.check_victory(second_player.token, new_board.board) if new_game.turn.even?
+
+    if win
+      new_board.display_board
+      puts "#{first_player.name} WINS THE GAME!!!!!!! CONGRATULATIONS" if new_game.turn.odd?
+      puts "#{second_player.name} WINS THE GAME!!!!!!! CONGRATULATIONS" if new_game.turn.even?
+      break
     end
-    puts 'NOBODY WON IT IS A DRAW'
+    #finish_game if win
+    puts 'NOBODY WON IT IS A DRAW' if new_game.turn == 9
   end
+ 
+  print "\nDo you want to play again? [y/n]: "
+  input = gets.chomp.downcase
 
-  def finish_game
-    @board.display_board
-    puts "#{@player1.name} WINS THE GAME!!!!!!! CONGRATULATIONS" if @turn.odd?
-    puts "#{@player2.name} WINS THE GAME!!!!!!! CONGRATULATIONS" if @turn.even?
-    exit(true)
-  end
+  break if input == 'n'
 
-  def check_victory(token)
-    counters = [0, 0, 0, 0] # [cols,rows,diag1,diag2]
-    (0..2).each do |i|
-      counters[0] = [board[i], board[i + 3], board[i + 6]].count(token) # Checks the positions for the columns
-      counters[1] = [board[i * 3], board[(i * 3) + 1], board[(i * 3) + 2]].count(token)
-      counters[2] += 1 if board[i * 4] == token # Counts if a position in the main diagonal
-      counters[3] += 1 if board[(i * 2) + 2] == token # Counts if a position in the oposite diagonal
-      return true if counters.include? 3
-    end
-    false
-  end
-
-  def check_win_move(token)
-    diag1 = []
-    diag2 = []
-
-    (0..2).each do |i|
-      col = [board[i], board[i + 3], board[i + 6]] # Gets the whole column
-      row = [board[i * 3], board[(i * 3) + 1], board[(i * 3) + 2]] # Gets the whole row
-      diag1.push(board[i * 4]) # Starts constructing the main diagonal
-      diag2.push(board[(i * 2) + 2]) # Starts constructing the oposite diagonal
-      return true if col.count(token) == 2 && col.any?(Numeric)
-
-      return true if row.count(token) == 2 && row.any?(Numeric)
-    end
-    return true if diag1.count(token) == 2 && diag1.any?(Numeric)
-
-    return true if diag2.count(token) == 2 && diag2.any?(Numeric)
-
-    false
-  end
 end
-
-class Player
-  attr_reader :name, :token
-  def initialize(name, token)
-    @name = name
-    @token = token
-  end
-end
-
-class Board < Array
-  def display_board
-    print "\n\t---------------\n\t"
-    each_with_index do |item, i|
-      print "| #{item} |"
-      print "\n\t---------------\n\t" if ((i + 1) % 3).zero?
-    end
-    print "\n"
-  end
-end
-
-# rubocop:enable Metrics/PerceivedComplexity,Metrics/CyclomaticComplexity
-
-new_board = Board[1, 2, 3, 4, 5, 6, 7, 8, 9]
-
-puts 'First player? '
-first_player = Player.new(gets.chomp.capitalize, 'X')
-puts 'Second player? '
-second_player = Player.new(gets.chomp.capitalize, 'O')
-
-new_game = Game.new(first_player, second_player, new_board)
-
-new_game.start_game
