+++ lib/player.rb
@@ -1,0 +1,7 @@
+class Player
+  attr_reader :name, :token
+  def initialize(name, token)
+    @name = name
+    @token = token
+  end
+end
